Implement "has_leading_newline"
This change adds a new TermOption, "has_leading_newline".

When this option is set to false (default is true),
TermLock::ensure_height() does not add a leading newline
(=the separator between the original line contents and the new
contents) to the output anymore.

When it is set to true, there is no change in ensure_height()'s
behavior (the leading newline is present).